AI/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens
attrs==23.2.0
audioread==3.0.1
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
boto3==1.34.138
botocore==1.34.138
build==1.2.1
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
cloudpathlib==0.16.0
cmake==3.29.6
colorama
coloredlogs==15.0.1
comm
confection==0.1.5
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.6
debugpy
decorator==4.4.2
Deprecated==1.2.14
distro==1.9.0
dlib
exceptiongroup
executing
fastapi==0.110.3
filelock==3.14.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
google-auth==2.29.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.64.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_metadata
importlib_resources==6.4.0
imutils==0.5.4
intel-openmp==2021.4.0
ipykernel
ipython
itsdangerous==2.2.0
jedi
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client
jupyter_core
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.2.5
langchain-openai==0.1.8
langchain-text-splitters==0.0.2
langcodes==3.4.0
langsmith==0.1.83
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.41.1
marisa-trie==1.1.1
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.7.5
matplotlib-inline
mkl==2021.4.0
mmh3==4.1.0
monotonic==1.6
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nest_asyncio
networkx==3.3
numba==0.58.1
numpy==1.24.4
oauthlib==3.2.2
onnxruntime==1.18.0
openai==1.35.7
opencv-python==4.10.0.84
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.3
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parso
pickleshare
pillow==10.3.0
platformdirs
pooch==1.8.2
posthog==3.5.0
preshed==3.0.9
proglog==0.1.10
prompt-toolkit
protobuf==4.25.3
psutil
pure-eval
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
Pygments
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
python-dateutil
python-dotenv==1.0.1
pytz==2024.1
pywin32==305.1
PyYAML==6.0.1
pyzmq
regex==2024.5.15
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.14.0
seaborn==0.12.2
six
smart-open==6.4.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
sqlparse==0.5.0
srsly==2.4.8
stack-data
starlette==0.37.2
sympy==1.12.1
tbb==2021.13.0
tenacity==8.3.0
thinc==8.2.3
threadpoolctl==3.5.0
tifffile==2024.7.2
tiktoken==0.7.0
tokenizers==0.19.1
tornado
tqdm==4.66.4
traitlets
transformers==4.42.3
typer==0.9.4
typing-inspect==0.9.0
typing_extensions
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.30.1
wasabi==1.1.3
watchfiles==0.22.0
wcwidth
weasel==0.3.4
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wikipedia==1.4.0
wrapt==1.16.0
yarl==1.9.4
zipp
